+++ src/base/window_manager.c
@@ -119,6 +119,10 @@ static ret_t window_manager_dispatch_window_event(widget_t* window, event_type_t
     window_manager_dispatch_top_window_changed(window->parent);
   }
 
+  if (type == EVT_WINDOW_TO_FOREGROUND) {
+    window->parent->key_target = window;
+  }
+
   return widget_dispatch(window->parent, (event_t*)&(evt));
 }
 
@@ -327,6 +331,10 @@ static ret_t window_manager_check_if_need_open_animation(const idle_info_t* info
   window_manager_dispatch_window_event(curr_win, EVT_WINDOW_WILL_OPEN);
 
   if (window_manager_create_animator(wm, curr_win, TRUE) != RET_OK) {
+    widget_t* prev_win = window_manager_find_prev_window(WIDGET(wm));
+    if (prev_win != NULL) {
+      window_manager_dispatch_window_event(prev_win, EVT_WINDOW_TO_BACKGROUND);
+    }
     window_manager_dispatch_window_event(curr_win, EVT_WINDOW_OPEN);
     widget_add_timer(curr_win, on_idle_invalidate, 100);
   }
@@ -461,6 +469,10 @@ ret_t window_manager_close_window(widget_t* widget, widget_t* window) {
 
   window_manager_dispatch_window_event(window, EVT_WINDOW_CLOSE);
   if (window_manager_check_if_need_close_animation(wm, window) != RET_OK) {
+    widget_t* prev_win = window_manager_find_prev_window(WIDGET(wm));
+    if (prev_win != NULL) {
+      window_manager_dispatch_window_event(prev_win, EVT_WINDOW_TO_FOREGROUND);
+    }
     widget_remove_child(widget, window);
     idle_add(window_manager_idle_destroy_window, window);
   }
@@ -613,6 +625,7 @@ static ret_t window_manager_paint_animation(widget_t* widget, canvas_t* c) {
 
   if (ret == RET_DONE) {
     bool_t is_open = wm->animator->open;
+    widget_t* prev_win = wm->animator->prev_win;
     widget_t* curr_win = wm->animator->curr_win;
     window_animator_destroy(wm->animator);
 
@@ -621,7 +634,10 @@ static ret_t window_manager_paint_animation(widget_t* widget, canvas_t* c) {
     wm->ignore_user_input = FALSE;
 
     if (is_open) {
+      window_manager_dispatch_window_event(prev_win, EVT_WINDOW_TO_BACKGROUND);
       window_manager_dispatch_window_event(curr_win, EVT_WINDOW_OPEN);
+    } else {
+      window_manager_dispatch_window_event(prev_win, EVT_WINDOW_TO_FOREGROUND);
     }
 
     if (wm->pending_close_window != NULL) {