+++ docs/changes.md
@@ -1,4 +1,8 @@
 # 最新动态
+* 2019/05/27
+    * 窗口在切换到前台和后台时分别触发事件EVT_WINDOW_TO_FOREGROUND和EVT_WINDOW_TO_BACKGROUND
+    * 窗口被切换到前台时，把窗口管理器的key_target设置到当前窗口。
+
 * 2019/05/25
   * 修改窗口中存在pages/slideview时焦点切换的问题。